backend/registerSubscription/requirements.txt
flask>=2.1.2
flask-cors>=3.0.10
google-cloud-storage>=2.4.0